Account class set up
next session try to make it so that the file runs and i can test the methods inside of it. Then create some test accounts to see if it will work. Then possibly do it so the user can input there own handle and description. adam is currently working on exception handling for the create a class as well.